Process VL6180 trame from serial and update test case

+++ core/include/WestBot/RobotRock/Vl6180x.hpp
@@ -26,7 +26,7 @@ public:
      *
      * \return Distance in mm.
      */
-    uint32_t distance() const;
+    uint8_t distance() const;
 
 private slots:
     void readData();
